moved stuff around. TODO get features to not be off
