+++ app/page.tsx
@@ -1,12 +1,12 @@
 'use client'
 
-import ExperimentPage from "@/components/ExperimentPage";
+import ExperimentPage from "@/app/components/ExperimentPage";
 import GrowthbookContainer from "@/app/components/GrowthbookContainer";
 
 /**
  * This is main page.
  * 
- * It ensures that growthbook in intialized before rendering the experiment page
+ * It ensures that growthbook is intialized before rendering the experiment page
  */
 export default function Home() {  
   return (
